requirements.txt
numpy==1.19.1
scipy==1.5.0
pandas==0.25.1
#torch==1.12.1+cu113
#torchvision==0.13.1+cu113
facenet-pytorch==0.4.0
opencv-python==4.4.0.44
Pillow==7.2.0
face-recognition==1.2.3
scikit-learn==0.22.2.post1
tqdm==4.46.0
pynvml==8.0.4
nvidia-ml-py3==7.352.0
tensorboard==2.2.1
tensorboardX==2.1
protobuf==3.20.3